selection sort algorithm implemented
