+++ application/src/test/java/com/clean_architecture_utils/algorithms/sorting/selection/RudeSelectionAlgorithmTest.java
@@ -1,18 +1,18 @@
-package com.clean_architecture_utils.algorithms.sorting.bubble;
+package com.clean_architecture_utils.algorithms.sorting.selection;
 
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 
-public class RudeBubbleSortingTest {
+public class RudeSelectionAlgorithmTest {
 
-    RudeBubbleSorting rudeBubbleSorting = new RudeBubbleSorting();
+    RudeSelectionSort rudeSelectionSort = new RudeSelectionSort();
 
     @Test
-    public void sortAnArrayWithTwoElements() {
+    public void sortArrayWithTwoElements(){
         int[] disorderedArray = {2, 1};
         int[] orderedArray = {1, 2};
-        rudeBubbleSorting.sort(disorderedArray);
+        rudeSelectionSort.sort(disorderedArray);
         assert Arrays.equals(disorderedArray, orderedArray);
     }
 
@@ -20,7 +20,7 @@ public class RudeBubbleSortingTest {
     public void sortAnArrayWithThreeElements() {
         int[] disorderedArray = {2, 1, 4};
         int[] orderedArray = {1, 2, 4};
-        rudeBubbleSorting.sort(disorderedArray);
+        rudeSelectionSort.sort(disorderedArray);
         assert Arrays.equals(disorderedArray, orderedArray);
     }
 
@@ -28,7 +28,7 @@ public class RudeBubbleSortingTest {
     public void sortAnArrayWithTenElements() {
         int[] disorderedArray = {2, 1, 4, 15, 900, 1999, 0, 99, 19};
         int[] orderedArray = {0, 1, 2, 4, 15, 19, 99, 900, 1999};
-        rudeBubbleSorting.sort(disorderedArray);
+        rudeSelectionSort.sort(disorderedArray);
         assert Arrays.equals(disorderedArray, orderedArray);
     }
 }
